ValidatorCLI - create a dedicated class to invoke validator-cli.jar (refs #15)
